# Monitor processes, unsubscribe them when they go down