* Use Process.monitor to monitor recipients
* Handle DOWN messages from recipients. Unsub them from NATS, remove from state.
* Process.demonitor whenever a receipient usubs

Addresses https://github.com/nats-io/nats.ex/issues/33

Let me know if this looks good, doing some testing on it now!