@acco 2024-02-19
Hey @mmmries – thanks for your work on this project! Just an update: been running this change in prod since I opened the PR and it seems to do the trick