@mmmries 2024-02-19
Hey @acco thanks so much for the submissions 🙌 

I'm thinking we should ask @autodidaddict to check this out as well as he's been helping out with the Elixir client and has a lot more context from within Synadia. Did you run any benchmarks on this by any chance? I was going back and forth on whether it would be best to handle exceptions when attempting to deliver messages vs actively monitoring client processes.