@acco 2024-02-19
Hey @mmmries – my pleasure!

I have not run any benchmarks on this.