@mmmries 2024-02-22
I ran a few benchmarks on this and found that things all looked good. I ran the `bench/request.exs` single-core performance test that does a roundtrip call and I also ran the `bench/server.exs` + `bench/client.exs` combined which goes between multiple VM instances with lots of parallelism. I ran this with 128 requesters and 5000 messages per requester

### main branch

```
mries@mries-ltmtqyn nats.ex % mix run bench/client.exs
It took 8.465676sec
	75599.39690581118 req/sec
	2362.4811533065995 kb/sec
	min: 103µs
	max: 15433µs
	median: 1444µs
	90th percentile: 2402µs
	average: 1575µs
	103,15433,1444,2402,1575,75599.39690581118,2362.4811533065995
```

### This PR branch

```
mries@mries-ltmtqyn nats.ex % mix run bench/client.exs
It took 8.452638sec
	75716.00723939674 req/sec
	2366.125226231148 kb/sec
	min: 92µs
	max: 14753µs
	median: 1440µs
	90th percentile: 2415µs
	average: 1575µs
	92,14753,1440,2415,1575,75716.00723939674,2366.125226231148
```

Ran this on erlang 25.1 and elixir 1.14.0 on a macbook pro. Seeing no significant different in performance which is amazing. I ❤️💙💜💚💛🧡🤎🖤 the Erlang VM for giving us these kinds of powerful tools