@mmmries 2024-02-22
Released as `gnat` version `1.8.1` https://hex.pm/packages/gnat/1.8.1

Thanks @acco for the contribution