@autodidaddict 2024-02-22
Can we get a github release for this?

On Thu, Feb 22, 2024 at 1:11 PM Michael ***@***.***> wrote:

> Released as gnat version 1.8.1 https://hex.pm/packages/gnat/1.8.1
>
> Thanks @acco <https://github.com/acco> for the contribution
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/nats-io/nats.ex/pull/149#issuecomment-1959997978>, or
> unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ADI4TINS6X643LWM2F6XH5DYU6C3PAVCNFSM6AAAAABDJGZ6DWVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMYTSNJZHE4TOOJXHA>
> .
> You are receiving this because you were mentioned.Message ID:
> ***@***.***>
>