@mmmries 2024-02-22
Yes, thanks for the reminder @autodidaddict

I created release 1.8.1 in github which also contains all of the jetstream integration work you did in version 1.8